Separate out entities in directories
